accept 64-bit offsets, for now

+++ src/wasm-s-parser.h
@@ -794,7 +794,7 @@ private:
       if (str[0] == 'a') {
         ret->align = atoi(eq);
       } else if (str[0] == 'o') {
-        ret->offset = atoi(eq);
+        ret->offset = atol(eq); // XXX https://github.com/WebAssembly/spec/issues/161
       } else onError();
       i++;
     }
